Kernel cleanup and more
Kernel cleanup, delete test files and upload Makefile and kernel.ld
