--- aquaOS/Version 0.5/bootloader/asd
@@ -1,1 +1,0 @@
-
