--- aquaOS/Version 0.5/kernel/as
@@ -1,1 +1,0 @@
-
