+++ aquaOS/Version 0.5/kernel/kernel.cpp
@@ -1,25 +1,4 @@
-#include <stdint.h>
-#include <stddef.h>
-#include "BasicRenderer.h"
-#include "cstr.h"
-#include "Framebuffer.h"
-#include "efiMemory.h"
-#include "memory.h"
-
-struct BootInfo
-{
-	Framebuffer* framebuffer;
-	PSF1_FONT* ps1_Font;
-	EFI_MEMORY_DESCRIPTOR* mMap;
-	uint64_t mMapSize;
-	uint64_t mMapDescSize;
-
-	uint64_t mReso;
-	uint64_t mResoSize;
-	uint64_t mResoDescSize;
-
-} ;
-
+#include "kernelUtil.h"
 
 /*/
 void Info()
@@ -63,5 +42,8 @@ extern "C" void _start(BootInfo* bootInfo)
 	}
 
 
-    return ;
+    while (true)
+	{
+		asm ("hlt");
+	}
 }
